Match extract_authorization expectations
`dwolla.rb#extract_authorization` expects oauthToken (which comes from `token`) to be `false`, `true`, or an actual `Oauth Token`. Defaulting to `nil` causes the test `oauthToken.is_a?(TrueClass)` to be false which causes `nil` to be used as the actual token. 

This change maintains pre-3.0.0 functionality based on setting `Dwolla::token`